requirements.txt
agora-token-builder==1.0.0
asgiref==3.4.1
Django==4.0.1
sqlparse==0.4.2
tzdata==2021.5
openai
bing_image_downloader
pywhatkit
googletrans==3.1.0a0
pyttsx3
pydub
langdetect
fpdf
django-widget-tweaks
pdf2docx
pytesseract
docx 
python-docx 
Pillow
PyPDF2
django-downloadview
pytube
tabula
pandas
openpyxl
googlesearch-python
xlwt
htmlmin
xhtml2pdf
pdfkit
docx2pdf
nltk
django-import-export
